app/shopify-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
psycopg = {extras = ["binary"], version = "^3.2.6"}
pydantic = "^2.11.2"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
stripe = "^12.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
